% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_chop_equal.R
\name{airdas_chop_equal}
\alias{airdas_chop_equal}
\title{Chop AirDAS effort segment length}
\usage{
airdas_chop_equal(x, seg.km, randpicks = NULL)
}
\arguments{
\item{x}{data frame; processed aerial DAS data from \code{\link{airdas_effort}}.
This data must contain exactly one continuous effort section, 
i.e. it must begin with a "T"/"R" event and end with a "E"/"O" event 
(with no "T"/"R"/"E"/"O" events in between)}

\item{seg.km}{numeric; length of effort segments. Passed from \code{\link{airdas_effort}}}

\item{randpicks}{numeric; segment positions that get leftover segment bits as described in Details}
}
\value{
List of:
\itemize{
  \item The AirDAS data, with a \code{effort_seg} column added..
  \item
  \item The index of 
}
}
\description{
Determine lengths of segments into which to chop aerial DAS effort data
}
\details{
This function chops continuous effort sections (henceforth 'effort sections') 
  from the processed AirDAS data into modeling segments (henceforth 'segments') of equal length. 
  Each effort sections runs from a T/R event to its corresponding E/O event. 

  If the extra length remaining after chopping is greater than or equal to half of the 
  target segment length (i.e. \code{>= 0.5*seg.km}), 
  it is randomly assigned to a portion of the effort section and 
  the associated distance provided as an offset value (Fig. 1a). 
  If the extra length is less than half of the target segment length (i.e. \code{< 0.5*seg.km}), 
  it is randomly added to one of the equal-length segments and the associated distance provided 
  as an offset value (Fig. 1b). 
  Therefore, the length of each segment is constrained to be between one half and one and one half of 
  \code{seg.km} (i.e. \code{0.5*seg.km to 1.5*seg.km}), 
  and the central tendency is approximately equal to the target segment length. 
  The only exception is when a continuous effort section is less than one half of the 
  target segment length (i.e. \code{< 0.5*seg.km}), 
  in which case the entire continuous effort section is a single segment (Fig. 1c).
  
  If the length of the effort section is zero and there are more than two events 
  (i.e., events between the "T"/R" and "E"/"O" events),
  the function throws a warning (but still returns a length of zero for that segment).
  
  If the column \code{dist_from_prev} does not exist 
  (it should be calculated in  \code{\link{airdas_effort}}), 
  then the distance between the lat/lon points of subsequent events 
  is calculated using \code{\link[swfscMisc]{distance}}, \code{method = "vincenty"}.
  
  randpicks description
  
  If segment length is 0, nothing..? 
  
  If segment length is less than target length, only make one segment
  
  If segment length is greater than the target length and remainder is greater than or equal to half of the target length, 
  the remainder is its own (randomly placed) segment
  
  If segment length is greater than the target length and remainder is less than half of the target length, 
  the remainder added to a random segment
}
\keyword{internal}
