% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_effort_chop.R
\name{airdas_effort_chop}
\alias{airdas_effort_chop}
\title{Chop AirDAS effort segment length}
\usage{
airdas_effort_chop(das.df, seg.km, randpicks = NULL)
}
\arguments{
\item{das.df}{data frame; processed aerial DAS data output from \link{airdas_process} (or subsequent airdas function)}

\item{seg.km}{numeric; length of effort segments}

\item{randpicks}{numeric; if not NULL, segment positions that get leftover segment bits todo}
}
\value{
List of...
}
\description{
Determine lengths of segments into which to chop aerial DAS effort data
}
\details{
Chop continuous effort section into effort segments (modeling segments)
  
  Calculates dist_from_prev using swfscMisc::distance, if necessary
  
  randpicks description
  
  If segment length is 0, nothing..? 
  
  If segment length is less than target length, only make one segment
  
  If segment length is greater than the target length and remainder is greater than or equal to half of the target length, 
  the remainder is its own (randomly placed) segment
  
  If segment length is greater than the target length and remainder is less than half of the target length, 
  the remainder added to a random segment
}
\examples{
d.proc <- airdas_process(system.file("airdas_sample.das", package = "swfscAirDAS"))
# airdas_effort_chop(d.proc, 1)

}
