% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_effort_segdata.R
\name{airdas_effort_segdata}
\alias{airdas_effort_segdata}
\title{Summarize data within effort segment}
\usage{
airdas_effort_segdata(x, subseg.lengths, eff.id)
}
\arguments{
\item{x}{data frame; todo}

\item{subseg.lengths}{todo}

\item{eff.id}{numeric; todo}
}
\value{
Data frame...
}
\description{
Summarize data for continuous effort section by effort segment
}
\details{
Get summary data for effort segment
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

y.chop <- airdas_effort_chop(y.proc[1:33, ], 1)
airdas_effort_segdata(y.chop[[1]], y.chop[[2]], 1)

}
