% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_format_copy.R
\name{airdas_format_copy}
\alias{airdas_format_copy}
\title{Aerial DAS format requirements}
\usage{
airdas_format_copy(file, ...)
}
\arguments{
\item{file}{character vector, the name of the file where the format pdf will be saved.
This argument is passed top the \code{to} argument of \link[base]{file.copy}}

\item{...}{arguments passed to \link[base]{file.copy}, such as \code{overwrite}}
}
\value{
\code{TRUE} is writing of file was successful, and \code{FALSE} otherwise
}
\description{
Save pdf document describing the required format of aerial DAS data
}
\details{
A wrapper function for \link[base]{file.copy}; 
  saves the pdf document describing the aerial DAS data format requirements.
  This function copies the pdf document located at \code{system.file("AirDAS_Format.pdf", package = "swfscAirDAS")}
  to \code{file}
}
\examples{
\dontrun{
airdas.format.pdf <- system.file("AirDAS_Format.pdf", package = "swfscAirDAS")
file.copy(from = airdas.format.pdf, to ="AirDAS_Format.pdf", overwrite = FALSE)
}

}
