% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_segdata_avg.R
\name{airdas_segdata_avg}
\alias{airdas_segdata_avg}
\title{Summarize data within effort segment}
\usage{
airdas_segdata_avg(x, subseg.lengths, eff.id)
}
\arguments{
\item{x}{data frame; todo}

\item{subseg.lengths}{todo}

\item{eff.id}{numeric; todo}
}
\value{
Data frame...
}
\description{
Summarize data for continuous effort section by effort segment
}
\details{
Get summary data for effort segment
}
\keyword{internal}
