% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_airdas_dfr.R
\name{as_airdas_dfr}
\alias{as_airdas_dfr}
\alias{as_airdas_dfr.airdas_dfr}
\alias{as_airdas_dfr.data.frame}
\title{Coerce object to a airdas_dfr object}
\usage{
as_airdas_dfr(x)

\method{as_airdas_dfr}{airdas_dfr}(x)

\method{as_airdas_dfr}{data.frame}(x)
}
\arguments{
\item{x}{An object to be coerced to class \code{airdas_dfr}}
}
\value{
An object of class \code{airdas_dfr}
}
\description{
Check if an object is of class \code{\link{airdas_dfr}}, or coerce it if possible.
}
\details{
Currently only data frames can be coerced to an object of class \code{\link{airdas_dfr}}.
If \code{x} does not have column names and classes as specified in \code{\link{airdas_dfr}},
then the function returns an error message detailing the first column that does not
meet the \code{\link{airdas_dfr}} requirements.
}
\seealso{
\code{\link{airdas_dfr-class}}
}
