% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_check.R
\name{airdas_check}
\alias{airdas_check}
\title{Check AirDAS file}
\usage{
airdas_check(file, file.type = "turtle", skip = 0, file.out = NULL)
}
\arguments{
\item{file}{filename(s) of one or more AirDAS files}

\item{file.type}{character; indicates the program used to create \code{file}.
Must be one of: "turtle", "caretta", "survey", or "phocoena" (case sensitive).
Default is "turtle". Passed to \code{\link{airdas_read}}}

\item{skip}{integer: see \code{\link[readr]{read_fwf}}. Default is 0. 
Passed to \code{\link{airdas_read}}}

\item{file.out}{filename to which to write the error log; 
default is \code{NULL}}
}
\value{
A data frame with five columns that list information about errors found 
in the AirDAS files: the file name, line number, 
index (row number) from the \code{airdas_read(file)} data frame, 
'ID' (pre-Data# columns from the DAS file), and description of the issue

If \code{file.out} is not \code{NULL}, then the error log is also
written to a text/csv file
}
\description{
Check that AirDAS file has accepted formatting and values
}
\details{
Checks that the following is true:
\itemize{
  \item Event codes are one of the following: 
    #, *, 1, A, C, E, O, P, R, s, S, t, T, V, W
  \item The effort dot matches effort determined using T, R, O, and E events
  \item All Data# columns, for events other than C events, are right-justified
  \item All * events have NA (blank) Data# columns
  \item Event/column pairs meet the following requirements:
}

\tabular{ll}{
  \emph{Item} \tab \emph{Requirement}\cr
  Viewing conditions \tab Must be one of: e, g, p, o, or NA (blank). Not case sensitive\cr
  Altitude           \tab Can be converted to a numeric value, and is not NA\cr
  Speed              \tab Can be converted to a numeric value, and is not NA\cr
  HKR              \tab Characters must all be one of: n, h, k, r, y, or NA (blank). Not case sensitive\cr
  Percent overcast \tab Can be converted to a numeric value, and is between 0 and 100\cr
  Beaufort         \tab Can be converted to a numeric value, and is between 0 and 9\cr
  Jellyfish        \tab Must be one of 0, 1, 2, 3, or NA (blank)\cr
  Horizontal sun   \tab Can be converted to a numeric value, and must be one of 0:12, or NA (blank)\cr
  Vertical sun     \tab Can be converted to a numeric value, and must be one of 0:4 or NA (blank)\cr
  Observers        \tab Each entry must be two characters\cr
  Observers        \tab No observer code can be used twice in the same P event\cr
  Sighting (mammal) \tab Angle and number of animals can be converted to a numeric value\cr
  Sighting (mammal) \tab Species code has exactly one or two characters\cr
  Sighting (mammal) \tab Observer code has exactly one or two characters\cr
  Sighting info     \tab Species percentages can be converted to a numeric value\cr
  Sighting info     \tab Unused columns of a '1' event must be NA (blank)\cr
  Resight           \tab Angle can be converted to a numeric value\cr
  Resight           \tab Unused resight columns must be NA (blank)\cr
  Turtle sighting   \tab Species code has exactly one or two characters\cr
  Turtle sighting   \tab Observer code has exactly one or two characters\cr
  Turtle sighting   \tab Angle and travel direction can be converted to a numeric value\cr
  Turtle sighting   \tab Tail visible must be one of y, n, u, or NA (blank). Case sensitive\cr
  Turtle sighting   \tab In TURTLE data, number of turtles can be converted to a numeric value\cr
  Turtle sighting   \tab In TURTLE data, turtle size can be converted to a numeric value\cr
  Turtle sighting   \tab In TURTLE data, the Data7 column must be NA (blank)\cr
}

Outstanding questions:
\itemize{
  \item Are there supposed to be equal numbers of T/O and R/E events?
  \item For which conditions should NA not be allowed?
  \item Check for valid fish ball/mola/jelly/crab pot codes in comments?
  \item Do checks for transect number?
}
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_check(y)

}
