% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_check.R
\name{airdas_check}
\alias{airdas_check}
\title{Check AirDAS file}
\usage{
airdas_check(file, file.out = NULL)
}
\arguments{
\item{file}{filename(s) of one or more AirDAS files}

\item{file.out}{filename to which to write the error log; 
default is \code{NULL}}
}
\value{
A data frame with five columns: the file name, line number, 
index (row number) from the \code{airdas_read(file)} data frame, 
'ID' (columns 4-39 from the DAS file), and description of the issue

If \code{file.out} is not \code{NULL}, then the error log is also
written to a text file
}
\description{
Check that AirDAS file has accepted formatting and values
}
\details{
Checks that the following is true:
\itemize{
  \item Event codes are one of the following: 
    #, *, 1, A, C, E, O, P, R, s, S, t, T, V, W
  \item The effort dot matches effort determined using T, R, O, and E events
  \item And event/column pairs meet the following requirements:
}

\tabular{llll}{
  \emph{Item} \tab \emph{Event} \tab \emph{Column} \tab \emph{Requirement}\cr
  Viewing conditions  \tab V \tab Data1-5 \tab Must be one of: e, g, p, o, or NA (blank)\cr
  Altitude            \tab A \tab Data1 \tab Can be converted to a numeric value\cr
  Speed               \tab A \tab Data2 \tab Can be converted to a numeric value\cr
  HKR              \tab W \tab Data1 \tab Characters must all be one of: n, h, k, r, or NA (blank)\cr
  Percent overcast \tab W \tab Data2 \tab Can be converted to a numeric value\cr
  Beaufort         \tab W \tab Data3 \tab Can be converted to a numeric value\cr
  Jellyfish        \tab W \tab Data4 \tab Must be one of 0, 1, 2, 3, or NA (blank)\cr
  Horizontal sun   \tab W \tab Data5 \tab Can be converted to a numeric value\cr
  Horizontal sun   \tab W \tab Data5 \tab Must be one of 0:12, or NA (blank)\cr
  Observers        \tab P \tab Data1-4 \tab Each entry must be two characters\cr
  NA    \tab S \tab Data3-4 \tab Can be converted to a numeric value\cr
  NA    \tab 1 \tab Data5-7 \tab Can be converted to a numeric value\cr
  Blank \tab 1 \tab Data1-4 \tab These columns must be NA (blank)\cr
  NA    \tab s \tab Data2   \tab Can be converted to a numeric value\cr
  Blank \tab s \tab Data3-7 \tab These columns must be NA (blank)\cr
  NA    \tab t \tab Data2, 4-5 \tab Can be converted to a numeric value\cr
  NA    \tab t \tab Data6 \tab Must be one of y, n, u, or NA (blank)\cr
  Blank \tab t \tab Data7 \tab This column must be NA (blank)\cr
}

Outstanding questions:
\itemize{
  \item Are there supposed to be equal numbers of T/O and R/E events?
  \item What else to add?
}
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_check(y)

}
