% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_check.R
\name{airdas_check}
\alias{airdas_check}
\title{Check AirDAS file}
\usage{
airdas_check(file, file.type = "turtle", skip = 0, file.out = NULL)
}
\arguments{
\item{file}{filename(s) of one or more AirDAS files}

\item{file.type}{character; indicates the program used to create \code{file}.
Must be one of: "turtle", "caretta", "survey", or "phocoena" (case sensitive).
Default is "turtle". Passed to \code{\link{airdas_read}}}

\item{skip}{integer: see \code{\link[readr]{read_fwf}}. Default is 0. 
Passed to \code{\link{airdas_read}}}

\item{file.out}{filename to which to write the error log; 
default is \code{NULL}}
}
\value{
A data frame with five columns that list information about errors found 
in the AirDAS files: the file name, line number, 
index (row number) from the \code{airdas_read(file)} data frame, 
'ID' (pre-Data# columns from the DAS file), and description of the issue. 
This data frame is sorted by the 'Description' column. 
If there are multiple issues with the same line, the issue descriptions
are concatenated together using \code{paste(..., collapse = "; ")}

If \code{file.out} is not \code{NULL}, then the error log is also
written to a text/csv file
}
\description{
Check that AirDAS file has accepted formatting and values
}
\details{
Checks that the following is true:
\itemize{
  \item Event codes are one of the following: 
    #, *, 1, A, C, E, O, P, R, s, S, t, T, V, W
  \item The effort dot matches effort determined using T, R, O, and E events
  \item There is an O event between each T event, and vice versa, i.e. 
    the data does not sart a new transect while still on a transect and
    does not end a transect when already not on a transect
  \item A T or R event does not occur while already on effort
  \item An E event does not occur while already off effort
  \item When the file ends, the data must be off effort and not still on a transect 
    (i.e. an O event must occurred more recently than a T event)
  \item All Data# columns for non-C events are right-justified
  \item The following events have NA (blank) Data# columns: *, R, E, O
  \item Event/column pairs meet the following requirements:
}

\tabular{ll}{
  \emph{Item} \tab \emph{Requirement}\cr
  Viewing conditions \tab Must be one of: e, g, p, o, or NA (blank). Not case sensitive\cr
  Altitude         \tab Can be converted to a numeric value, and is not NA\cr
  Speed            \tab Can be converted to a numeric value, and is not NA\cr
  HKR              \tab Characters must consist of: n, h, k, r, or NA (blank). 
    Not case sensitive; y is also accepted for PHOCOENA data\cr
  Percent overcast \tab Must be a whole number between 0 and 100\cr
  Beaufort         \tab Must be a whole number between 0 and 9\cr
  Jellyfish        \tab Must be one of 0, 1, 2, 3, or NA (blank)\cr
  Horizontal sun   \tab Must be one of 0:12, or NA (blank)\cr
  Vertical sun     \tab Must be one of 0:4 or NA (blank)\cr
  Observers        \tab Each entry must be two characters, and no observer code can be used twice in the same P event\cr
  Sighting (mammal) \tab Angle must be a whole number between -90 and 90\cr
  Sighting (mammal) \tab Group size must be a whole number between 1 and 5000\cr
  Sighting (mammal) \tab Species code(s) must be exactly two characters\cr
  Sighting (mammal) \tab Observer code must be exactly two characters\cr
  Sighting info     \tab Species percentages can be converted to numeric values, and sum to 100\cr
  Sighting info     \tab Unused columns (DateTime, Lat, Lon, and Data1-4) of a '1' event must be NA (blank)\cr
  Sighting info     \tab Every S code with multiple species has a 1 code immediately after it\cr
  Sighting info     \tab For every 1 code with n non-NA sighting percentages, 
    the event before is an S event with n non-NA species codes\cr
  Resight           \tab Angle can be converted to a numeric value\cr
  Resight           \tab Unused resight columns (Data3-7) must be NA (blank)\cr
  Turtle sighting   \tab Angle must be a whole number between -90 and 90\cr
  Turtle sighting   \tab Group size must be a whole number between 1 and 10 (only included in CARETTA data)\cr
  Turtle sighting   \tab Species code must be exactly two characters, and one of dc, cc, lo, cm, uh, ut, or lv\cr
  Turtle sighting   \tab Observer code must be exactly two characters\cr
  Turtle sighting   \tab Turtle size must be a whole number between 1 and 9; 
    one of s, m, l is also accepted for CARETTA data\cr
  Turtle sighting   \tab Travel direction must be a whole number between 0 and 360 (only included in TURTLE data)\cr
  Turtle sighting   \tab Tail visible must be one of y, n, u, or NA (blank). Case sensitive\cr
  Turtle sighting   \tab In TURTLE data, the Data7 column must be NA (blank)\cr
}

Outstanding questions/todo:
\itemize{
  \item For which conditions should NA not be allowed?
  \item For which sighting data bits should NAs not be allowed (e.g. Obs, angle, sp, Gs)?
  \item Check for valid fish ball/mola/jelly/crab pot codes in comments?
  \item Do checks for transect number?
  \item Should description items be grouped together by line number?
  \item todo: Check that sighting observer is one of currently listed observers
  \item todo: Check that sighting observer side matches angle +/-
  \item todo: some way to load spcodes.dat to check species codes?
  \item todo: Check for equal number of non-NA species codes and percentages (including 1 sp code for no 1 event)
}
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_check(y)

}
