% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_chop_condition.R
\name{airdas_chop_condition}
\alias{airdas_chop_condition}
\alias{airdas_chop_condition.data.frame}
\alias{airdas_chop_condition.airdas_df}
\title{Chop AirDAS data - condition}
\usage{
airdas_chop_condition(x, ...)

\method{airdas_chop_condition}{data.frame}(x, ...)

\method{airdas_chop_condition}{airdas_df}(x, seg.km.min = 0.1, ...)
}
\arguments{
\item{x}{\code{airdas_df} object, 
or a data frame that can be coerced to a \code{airdas_df} object. 
This data must be filtered for 'OnEffort' events; 
see the Details section below}

\item{...}{ignored}

\item{seg.km.min}{numeric; all segments shorter than this distance, 
which is in kilometers,  will be smushed with the segment 
immediately following them. 
Default is 0.1}
}
\value{
List?
}
\description{
Chop AirDAS data into a new effort segment every time any condition changes
}
\details{
This function is intended to be called by \code{\link{airdas_effort}} 
  when the "condition" method is specified. 
  Thus, \code{x} must be filtered for events (rows) where either
  the 'OnEffort' column is \code{TRUE} or the 'Event' column is "E" or "O"; 
  see \code{\link{airdas_effort}} for more details. 
  Continuous effort sections (henceforth 'effort sections') from \code{x}
  are chopped into modeling segments (henceforth 'segments') every time a condition changes 
  Each effort section runs from a T/R event to its corresponding E/O event. 
  Then \code{\link{airdas_segdata_avg}} is called to get relevant segdata information.
  
  The following conditions trigger the creation of a new segment:
  viewing conditions, altitude, speed, HKR, percent overcast (cloud cover), 
  Beaufort, Jellyfish code, and horizontal sun
  
  All segments with length zero will be smushed with the segment immediately following them.
  This will take care of event series that all have the same location, such as TVPAW.
  
  TODO
}
\keyword{internal}
