% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_effort.R
\name{airdas_effort}
\alias{airdas_effort}
\alias{airdas_effort.data.frame}
\alias{airdas_effort.airdas_df}
\title{Summarize aerial DAS effort}
\usage{
airdas_effort(x, ...)

\method{airdas_effort}{data.frame}(x, ...)

\method{airdas_effort}{airdas_df}(x, sp.codes, ...)
}
\arguments{
\item{x}{\code{airdas_df} object; output from \code{\link{airdas_process}}, 
or a data frame that can be coerced to a \code{airdas_df} object}

\item{...}{passed to TODO}

\item{sp.codes}{character; species code(s) to include in segdata}
}
\value{
List of three data frames: 
  \itemize{
    \item segdata: one row for every segment, and columns for information such as
      unique segment number, start/end/midpoint coordinates, conditions (e.g. Beaufort), 
      and number of sightings and number of animals on that segment for evey species 
      indicated in \code{sp.codes}. 
    \item siteinfo: details for each sighting, including the segdata segment number it belongs on, 
      segment mid points (lat/lon), and whether 
      the sighting was included in the segdata segments (column \code{included}), 
      in addition to the other output information described in \code{\link{airdas_sight}}.
    \item randpicks: see \code{\link{airdas_chop_equal}}
  }
}
\description{
Summarize aerial DAS data by effort segments
}
\details{
Chops processed AirDAS data into modeling segments (henceforth 'segments'), 
  and assigns sightings and related information (e.g., weather conditions) to each segment. 
  This function returns all relevant information for the effort segments and 
  associated sightings ('segdata' and 'siteinfo', respectively). 
  Before chopping, the AirDAS data is filtered for events (rows) where either
  the 'OnEffort' column is \code{TRUE} or the 'Event' column is "E" or "O". 
  In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'), 
  where effort sections run from "T"/"R" to "E"/"O" events (inclusive). 

  Currently, the only available chopping method is chopping effort sections into 
  equal-length modeling segments of length \code{seg.km}, 
  and doing a weighted average of the conditions for the length of that segment. 
  See \code{\link{airdas_chop_equal}} for more details about the segment chopping.

  The function \code{\link[swfscMisc]{distance}}, \code{method = "vincenty"}, is used to
  calculate the distance (in km) between the lat/lon points of subsequent events.
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")

y.proc <- airdas_process(y)
airdas_effort(y.proc, sp.codes = c("mn", "bm"), seg.km = 3)

y.rand <- system.file("airdas_sample_randpicks.csv", package = "swfscAirDAS")
airdas_effort(
  y.proc, sp.codes = c("mn", "bm"), seg.km = 3, randpicks.load = y.rand
)

}
