% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_effort.R
\name{airdas_effort}
\alias{airdas_effort}
\title{Summarize aerial DAS effort}
\usage{
airdas_effort(
  das.df,
  seg.km,
  sp.codes,
  randpicks.load = NULL,
  randpicks.save = NULL
)
}
\arguments{
\item{das.df}{data frame; processed aerial DAS data output from \link{airdas_process} (or subsequent airdas function)}

\item{seg.km}{numeric; length of effort segments}

\item{sp.codes}{character; species codes to include in segdata (and siteinfo?)}

\item{randpicks.load}{character; file to run read.csv on and pass output to airdas_effort_chop TODO}

\item{randpicks.save}{character; if not}
}
\value{
List of segdata, siteinfo, and randpicks data frames
}
\description{
Summarize aerial DAS data by effort segments
}
\details{
Todo

  "continuous effort sections" and "modeling segments"

  Filters for events where \code{OnEffort} is \code{TRUE}
  
  Continuous effort sections ID'd by T and R events

  Uses \link[swfscMisc]{distance}, \code{method = "vincenty"} to calculate distance (in km) between events (lat/lon points)
  
  Describe segdata, siteinfo, and randpicks
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")

y.proc <- airdas_process(y)
airdas_effort(y.proc, 3, sp.codes = c("mn", "bm"))

y.rand <- system.file("airdas_sample_randpicks.csv", package = "swfscAirDAS")
airdas_effort(
  y.proc, 3, sp.codes = c("mn", "bm"), randpicks.load = y.rand
)

}
