% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_effort.R
\name{airdas_effort}
\alias{airdas_effort}
\alias{airdas_effort.data.frame}
\alias{airdas_effort.airdas_df}
\title{Summarize AirDAS effort}
\usage{
airdas_effort(x, ...)

\method{airdas_effort}{data.frame}(x, ...)

\method{airdas_effort}{airdas_df}(x, method, sp.codes, dist.method = "vincenty", ...)
}
\arguments{
\item{x}{\code{airdas_df} object; output from \code{\link{airdas_process}}, 
or a data frame that can be coerced to a \code{airdas_df} object}

\item{...}{arguments passed to the chopping function specified using \code{method}}

\item{method}{character; method to use to chop AirDAS data into effort segments
Can be "equallength" or "condition" (case-sensitive) to use 
\code{\link{airdas_chop_equal}} or \code{\link{airdas_chop_condition}}, respectively}

\item{sp.codes}{character; species code(s) to include in segdata}

\item{dist.method}{character;
method to use to calculate distance between lat/lon coordinates.
Can be "greatcircle" to use the great circle distance method (TODO - add ref),
or one of "lawofcosines", "haversine", or "vincenty" to use
\code{\link[swfscMisc]{distance}}. Default is "vincenty"}
}
\value{
List of three data frames: 
  \itemize{
    \item segdata: one row for every segment, and columns for information including
      unique segment number, start/end/midpoint coordinates, conditions (e.g. Beaufort), 
      and number of sightings and number of animals on that segment for every species 
      indicated in \code{sp.codes}. 
    \item siteinfo: details for all sightings in \code{x}, including: 
      the unique segment number it is associated with, segment mid points (lat/lon), 
      and whether the sighting was included in the segdata counts (column \code{included}), 
      in addition to the other output information described in \code{\link{airdas_sight}}.
    \item randpicks: see \code{\link{airdas_chop_equal}}
  }
}
\description{
Chop AirDAS data into effort segments
}
\details{
This is the top-level function for chopping processed AirDAS data 
  into modeling segments (henceforth 'segments'), and assigning sightings 
  and related information (e.g., weather conditions) to each segment. 
  This function returns data frames with all relevant information for the 
  effort segments and associated sightings ('segdata' and 'siteinfo', respectively). 
  Before chopping, the AirDAS data is filtered for events (rows) where either
  the 'OnEffort' column is \code{TRUE} or the 'Event' column is "E" or "O". 
  In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'), 
  where effort sections run from "T"/"R" to "E"/"O" events (inclusive), 
  and then passed to the chopping function specified using \code{method}. 

  Currently, two chopping methods are available: "equallength" and "condition". 
  The "equallength" method consists of
  chopping effort sections into equal-length segments of length \code{seg.km}, 
  and doing a weighted average of the conditions for the length of that segment. 
  See \code{\link{airdas_chop_equal}} for more details about this method, 
  including arguments that must be passed to it via \code{...}.
  
  When using the "condition" method, effort sections are chopped into segments
  every time a condition changes, 
  therby ensuring that the conditions are consistent across the entire segment.
  See \code{\link{airdas_chop_condition}} for more details about this method, 
  including arguments that must be passed to it via \code{...}.
  
  The sightings included in the segdata counts are filtered as follows: 
  Beaufort is less than or equal to five, the declination angle is less than 78, 
  it is a standard sighting, and the sighting was made while on effort. 
  Siteinfo, which contains all sightings in \code{x} (see below), 
  contains a column ('included') that indicates whether or not the sighting was
  included in the segdata counts.
  
  All on effort events must not have \code{NA} Lat or Lon values; 
  remember that in \code{\link{airdas_process}}, S Lat/Lon values
  were passed to 1 events. 
  TODO: Should this function verbosely remove these events?

  The distance between the lat/lon points of subsequent events
  is calculated using the method specified in \code{dist.method}
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

# Using "equallength" method
y.rand <- system.file("airdas_sample_randpicks.csv", package = "swfscAirDAS")
airdas_effort(
  y.proc, method = "equallength", sp.codes = c("mn", "bm"), 
  seg.km = 3, randpicks.load = y.rand, num.cores = 1
)

# Using "condition" method
airdas_effort(
  y.proc, method = "condition", sp.codes = c("mn", "bm"), 
  seg.km.min = 0.05, num.cores = 1
)

}
