% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_effort_segdata.R
\name{airdas_effort_segdata}
\alias{airdas_effort_segdata}
\title{Summarize data within effort segment}
\usage{
airdas_effort_segdata(das.df, subseg.lengths, eff.id)
}
\arguments{
\item{das.df}{data frame; todo}

\item{subseg.lengths}{todo}

\item{eff.id}{numeric; todo}
}
\value{
Data frame...
}
\description{
Summarize data for continuous effort section by effort segment
}
\details{
Get summary data for effort segment
}
\examples{
#TODO

}
