% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_format_pdf.R
\name{airdas_format_pdf}
\alias{airdas_format_pdf}
\title{Aerial DAS format requirements}
\usage{
airdas_format_pdf(file, ...)
}
\arguments{
\item{file}{character vector, the name of the file where the pdf will be saved}

\item{...}{arguments passed to \code{\link[base:files]{file.copy}}, such as \code{overwrite}}
}
\value{
\code{TRUE} if writing of file was successful, and \code{FALSE} otherwise
}
\description{
Save the pdf document describing the aerial DAS format required by \code{swfscAirDAS} to a specified file
}
\details{
A wrapper function for \code{\link[base:files]{file.copy}}. This function
  saves the pdf document describing the aerial DAS data format requirements by 
  copying the pdf document located at \code{system.file("AirDAS_Format.pdf", package = "swfscAirDAS")}
  to \code{file}
}
\examples{
\dontrun{
airdas_format_pdf("AirDAS_Format.pdf", overwrite = FALSE)
}

}
