% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_process.R
\name{airdas_process}
\alias{airdas_process}
\alias{airdas_process.character}
\alias{airdas_process.data.frame}
\title{Process aerial survey DAS data}
\usage{
airdas_process(x, ...)

\method{airdas_process}{character}(x, ...)

\method{airdas_process}{data.frame}(x, days.gap = 0.5/24, ...)
}
\arguments{
\item{x}{Either a data frame (the output of \link{airdas_read}) or
a character which is then passed to \link{airdas_read}}

\item{...}{Ignore}

\item{days.gap}{numeric of length 1; time gap (in days) used to identify 
when propogated info (weather/env, observers, etc) is reset. 
The default is 30 minutes}
}
\value{
Processed aerial survey DAS data...
}
\description{
Process output of \link{airdas_read}
}
\details{
Read...
  Adapted from \code{\link[swfscMisc]{das.read}}
 TODO: describe columns created?
 TODO: any error checking?
}
\examples{
# TODO
# d <- do.call(rbind, lapply(list.files("../airdas/airDAS_files", full.names = TRUE), airdas_read))
# x <- airdas_process(d)

}
