% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_process.R
\name{airdas_process}
\alias{airdas_process}
\alias{airdas_process.character}
\alias{airdas_process.data.frame}
\alias{airdas_process.airdas_dfr}
\title{Process aerial survey DAS data}
\usage{
airdas_process(x, ...)

\method{airdas_process}{character}(x, ...)

\method{airdas_process}{data.frame}(x, ...)

\method{airdas_process}{airdas_dfr}(
  x,
  days.gap.part = 0.5/24,
  days.gap.full = 12/24,
  gap.message = FALSE,
  ...
)
}
\arguments{
\item{x}{either a data frame (the output of \code{\link{airdas_read}}), 
or a character (filepath) which is first passed to \code{\link{airdas_read}}}

\item{...}{ignored}

\item{days.gap.part}{numeric of length 1; 
time gap (in days) used to identify when a 'partial reset' is performed, 
i.e. when propogated info (weather, observers, etc) is reset. 
Default is 30 minutes; must be less than or equal to \code{days.gap.full}}

\item{days.gap.full}{numeric of length 1; 
time gap (in days) used to identify when a 'full reset; is performed, 
i.e. when all info (transect number and propogated info) is reset. 
Default is 12 hours}

\item{gap.message}{logical; should messages be printed detailing which row(s) of the 
output data frame were partially or fully reset?}
}
\value{
An \code{airdas_df} object, which is also a data frame.
  It consists of the input data frame, i.e. the output of \code{\link{airdas_read}},
  with the following columns added:
  \tabular{lll}{
    \emph{State/condition}            \tab \emph{Column name} \tab \emph{Data source}\cr
    On/off effort                     \tab OnEffort \tab T/R and O/E events\cr
    Transect code                     \tab Trans    \tab Event: T; Column: Data1\cr
    Beaufort sea state                \tab Bft      \tab Event: W; Column: Data3\cr
    Percent overcast (cloud cover)    \tab CCover   \tab Event: W; Column: Data2\cr
    Jellyfish code                    \tab Trans    \tab Event: W; Column: Data4\cr
    Horizontal sun (clock system)     \tab HorizSun \tab Event: W; Column: Data5\cr
    Haze/Kelp/Red tide code           \tab HKR      \tab Event: W; Column: Data1\cr
    Haze (from HKR code)              \tab Haze     \tab HKR\cr
    Kelp (from HKR code)              \tab Kelp     \tab HKR\cr
    Red tide (from HKR code)          \tab RedTide  \tab HKR\cr
    Left observer                     \tab ObsL     \tab Event: P; Column: Data1\cr
    Belly observer                    \tab ObsB     \tab Event: P; Column: Data2\cr
    Right observer                    \tab ObsR     \tab Event: P; Column: Data3\cr
    Data recorder                     \tab Rec      \tab Event: P; Column: Data4\cr
    Altitude (feet)                   \tab AltFt    \tab Event: A; Column: Data1\cr
    Speed (knots)                     \tab SpKnot   \tab Event: A; Column: Data2\cr
    Viewing condition - left inside   \tab VLI      \tab Event: V; Column: Data1\cr
    Viewing condition - left outside  \tab VLO      \tab Event: V; Column: Data2\cr
    Viewing condition - belly         \tab VB       \tab Event: V; Column: Data3\cr
    Viewing condition - right inside  \tab VRI      \tab Event: V; Column: Data4\cr
    Viewing condition - right outside \tab VRO      \tab Event: V; Column: Data5\cr
  }
}
\description{
Process AirDAS data (the output of \code{\link{airdas_read}}), 
  including extracting state and condition information for each AirDAS event
}
\details{
If \code{x} is a character, it is assumed to be a filepath 
  and first passed to \code{\link{airdas_read}}.
  This output is then passed to \code{airdas_process}.
     
  AirDAS data is event-based, meaning most events indicate when a state or weather condition changes.
  For instance, a 'W' event indicates when the Beaufort sea state changes, and 
  the Beaufort is the same for subsequent events until the next 'W' event.
  For each state/condition: a new column is created, 
  the state/condition information is extracted from relevant events, 
  and extracted information is propagated to appropriate subsequent rows (events). 
  Thus, each row in the output data frame contains all 
  pertinent state/condition information for that row.
  
  The following assumptions/decisions are made during processing:
  \itemize{
    \item All '#' events (deleted events) are removed
    \item 'Datetime', 'Lat', and 'Lon' information are added to '1' events
    \item Effort is determined as follows: T/R events turns effort on, 
      and O/E events turn effort off. 
      T/R events themselves will be on effort, while O/E events will be off effort.
      The 'EffortDot' column is ignored
    \item Missing values are \code{NA} rather than \code{-1}
  }
  
  See \code{\link{airdas_format_pdf}} for more information about AirDAS format requirements
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
airdas_process(y)

y.read <- airdas_read(y)
airdas_process(y.read)

}
