% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_process.R
\name{airdas_process}
\alias{airdas_process}
\title{Process aerial survey DAS data}
\usage{
airdas_process(das.df, days.gap = 0.5/24)
}
\arguments{
\item{das.df}{data frame; the output of \link{airdas_read}}

\item{days.gap}{numeric of length 1; time gap (in days) used to identify 
when propogated info (weather/env, observers, etc) is reset}
}
\value{
Processed aerial survey DAS data...
}
\description{
Process output of \link{airdas_read}
}
\details{
Read...
  Adapted from \code{\link[swfscMisc]{das.read}}
 TODO: describe columns created?
 TODO: any error checking?
}
\examples{
# TODO
# d <- do.call(rbind, lapply(list.files("../airdas/airDAS_files", full.names = TRUE), airdas_read))
# x <- airdas_process(d)

}
