% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_read.R
\name{airdas_read}
\alias{airdas_read}
\title{Read AirDAS data}
\usage{
airdas_read(file, tz = "UTC")
}
\arguments{
\item{file}{filename of an aerial survey DAS file}

\item{tz}{character; see \link[base]{strptime}. Default is UTC}
}
\value{
Data frame with AirDAS data parsed into columns.
  The columns are parsed as raw characters as described in 'Details', 
  with the following exceptions:
  \itemize{
    \item \code{EffortDot}: logical; \code{TRUE} if "." was present, \code{FALSE} otherwise
    \item \code{DateTime}: \code{POSIXct}; combination of 'Date' and 'Time' columns with timezone \code{tz}
    \item \code{Lat}: \code{numeric}; 'Latitude' column converted to decimal degrees in range [-90, 90]
    \item \code{Lon}: \code{numeric}; 'Longitude' column converted to decimal degrees in range [-180, 180]
    \item \code{Data#}: leading/trailing whitespace trimmed for non-comment events (rows where \code{Event} is not "C" or "c")
    \item \code{file_das}: character; filename (note not the full filepath) from \code{file} argument
    \item \code{event_num}: integer; 'Event number' as an integer
    \item \code{line_num}: integer; line number of each data row
  }
}
\description{
Read fixed-width aerial survey DAS text file generated by turtle4d,
  or associated program, and parse into data frame
}
\details{
Parses aerial survey DAS data into columns.
  The provided AirDAS file must adhere to the following column number and format specifications 
  (See the "AirDAS_format" vignette for more information):
  \tabular{lrr}{
    \emph{Item}  \tab \emph{Columns} \tab \emph{Format}\cr
    Event number \tab 1-3\cr
    Event        \tab 4\cr
    Effort dot   \tab 5\cr
    Time         \tab 6-11  \tab HHMMSS\cr
    Date         \tab 13-18 \tab MMDDYY\cr
    Latitude     \tab 20-28 \tab NDD:MM.MM\cr
    Longitude    \tab 30-39 \tab WDDD:MM.MM\cr
    Data1        \tab 40-44\cr
    Data2        \tab 45-49\cr
    Data3        \tab 50-54\cr
    Data4        \tab 55-59\cr
    Data5        \tab 60-64\cr
    Data6        \tab 65-69\cr
    Data7        \tab 70+\cr
  }
}
\examples{
airdas_read(system.file("airdas_sample.das", package = "swfscAirDAS"))

}
