% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_read.R
\name{airdas_read}
\alias{airdas_read}
\title{Read AirDAS file(s)}
\usage{
airdas_read(file, tz = "UTC")
}
\arguments{
\item{file}{filename(s) of one or more AirDAS files}

\item{tz}{character; see \link[base]{strptime}. Default is UTC}
}
\value{
An \code{airdas_dfr} object, which is also a data frame, with AirDAS data read into columns.
  The data are read into the data frame as characters as described in 'Details',
  with the following exceptions:
  \tabular{lll}{
    \emph{Name} \tab \emph{Class} \tab \emph{Details}\cr
    EffortDot \tab logical   \tab \code{TRUE} if "." was present, and \code{FALSE} otherwise\cr
    DateTime  \tab POSIXct   \tab combination of 'Date' and 'Time' columns with time zone \code{tz}\cr
    Lat       \tab numeric   \tab 'Latitude' column converted to decimal degrees in range [-90, 90]\cr
    Lon       \tab numeric   \tab 'Longitude' column converted to decimal degrees in range [-180, 180]\cr
    Data#     \tab character \tab leading/trailing whitespace trimmed for non-comment events (i.e. rows where 'Event' is not "C" or "c")\cr
    EventNum  \tab integer   \tab 'Event number' coerced to an integer using \code{\link[base:integer]{as.integer}}\cr
    file_das  \tab character \tab filename, extracted from the \code{file} argument using \code{\link[base]{basename}}\cr
    line_num  \tab integer   \tab line number of each data row\cr
  }
}
\description{
Read one or more fixed-width aerial survey DAS text file(s) 
  generated by TURTLEP (or an associated program) into a data frame, 
  where each line is data for a specific event
}
\details{
Reads/parses aerial survey DAS data into columns of a data frame.
  If \code{file} contains multiple filenames, then the individual 
  data frames will be concatenated.
  The provided AirDAS file must adhere to the following column number and format specifications: 
  (see the "AirDAS_format" vignette for more information):
  \tabular{lrr}{
    \emph{Item}  \tab \emph{Columns} \tab \emph{Format}\cr
    Event number \tab 1-3\cr
    Event        \tab 4\cr
    Effort dot   \tab 5\cr
    Time         \tab 6-11  \tab HHMMSS\cr
    Date         \tab 13-18 \tab MMDDYY\cr
    Latitude     \tab 20-28 \tab NDD:MM.MM\cr
    Longitude    \tab 30-39 \tab WDDD:MM.MM\cr
    Data1        \tab 40-44\cr
    Data2        \tab 45-49\cr
    Data3        \tab 50-54\cr
    Data4        \tab 55-59\cr
    Data5        \tab 60-64\cr
    Data6        \tab 65-69\cr
    Data7        \tab 70+\cr
  }
   
  See \code{\link{airdas_format_pdf}} for more information about AirDAS format requirements
}
\examples{
airdas_read(system.file("airdas_sample.das", package = "swfscAirDAS"))

}
