% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_segdata.R
\name{airdas_segdata}
\alias{airdas_segdata}
\alias{airdas_segdata.data.frame}
\alias{airdas_segdata.airdas_df}
\title{Summarize AirDAS data for a continuous effort section}
\usage{
airdas_segdata(x, ...)

\method{airdas_segdata}{data.frame}(x, ...)

\method{airdas_segdata}{airdas_df}(x, conditions, segdata.method, seg.lengths, section.id, ...)
}
\arguments{
\item{x}{\code{airdas_df} object, 
or a data frame that can be coerced to a \code{airdas_df} object.
Must must be filtered for 'OnEffort' events and 
contain a single continuous effort section of AirDAS data; 
see the Details section below}

\item{...}{ignored}

\item{conditions}{see \code{\link{airdas_effort}}, or
see Details section for more information}

\item{segdata.method}{character; either \code{"avg"} or \code{"maxdist"}.
\code{"avg"} means the condition values will be
calculated as a weighted average by distance, while
\code{"maxdist"} means the condition values will be those recorded
for the longest distance during that segment}

\item{seg.lengths}{numeric; length of the modeling segments 
into which \code{x} will be chopped}

\item{section.id}{numeric; the ID of \code{x} (the current continuous effort section)}
}
\value{
Data frame with the segdata information described above
  and in \code{\link{airdas_effort}}
}
\description{
Summarize AirDAS effort data by effort segment, while averaging conditions
}
\details{
This function should be called by one of the airdas_chop functions, 
  e.g. \code{\link{airdas_chop_equal}}, meaning
  users should work to avoid calling it themselves.
  It loops through the events in \code{x}, calculating and storing relevant
  information for each modeling segment as it goes. 
  Because \code{x} is a continuous effort section, it must begin with 
  a "T" or "R" event and end with the corresponding "E" or "O" event.

  For each segment, this function reports the 
  segment ID, transect code, the start/end/midpoints (lat/lon), segment length, 
  year, month, day, time, observers, and average conditions.
  
  The segment ID is designated as \code{eff_id} _ index of the modeling segment.
  Thus, if \code{section.id} is \code{1}, then the segment ID for 
  the second segment from \code{x} is \code{"1_2"}.
  
  The average condition values are calculated as a weighted average by distance, 
  and reported for the following:
  Beaufort, cloud cover, altitude in feet, aircraft speed in knots, jellyfish code, 
  and the percentage of the segment with each of haze, kelp, and red tide. 
  For logical columns such as Haze, the reported value is the percentage 
  (in decimals) of the segment in which that condition was \code{TRUE}.   
  Transect number and file name are also also incldued in the segdata output;
  these values are (should be) all consistent across the whole effort section,
  and thus across all segments in \code{x}.
  
  \code{\link[swfscMisc]{bearing}} and \code{\link[swfscMisc]{destination}}, 
  method = "vincenty",
  are used to help calculate the segment start, mid, and end points.
}
\seealso{
airdas_segdata_max
}
\keyword{internal}
