% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_segdata_avg.R
\name{airdas_segdata_avg}
\alias{airdas_segdata_avg}
\alias{airdas_segdata_avg.data.frame}
\alias{airdas_segdata_avg.airdas_df}
\title{Summarize data for a continuous effort section}
\usage{
airdas_segdata_avg(x, ...)

\method{airdas_segdata_avg}{data.frame}(x, ...)

\method{airdas_segdata_avg}{airdas_df}(x, seg.lengths, eff.id, ...)
}
\arguments{
\item{x}{\code{airdas_df} object, 
or a data frame that can be coerced to a \code{airdas_df} object.
Must must be filtered for 'OnEffort' events and 
contain a single continuous effort section of AirDAS data; 
see the Details section below}

\item{...}{ignored}

\item{seg.lengths}{numeric; length of the modeling segments 
into which \code{x} will be chopped}

\item{eff.id}{numeric; the ID of \code{x} (the current continuous effort section)}
}
\value{
Data frame with the segdata information described above
}
\description{
Summarize effort data by effort segment, while averaging conditions
}
\details{
This function should be called by \code{\link{airdas_chop_equal}};
  users should (nearly always) not call it themselves.
  It loops through the events in \code{x}, calculating and storing relevant
  information for each modeling segment as it goes. 
  Because \code{x} is a continuous effort section, it must begin with 
  a "T" or "R" event and end with the corresponding "E" or "O" event.

  For each segment, this function reports the 
  segment ID, transect code, the start/end/midpoints (lat/lon), segment length, 
  year, month, day, time, observers, and average conditions.
  
  The segment ID is designated as \code{eff_id} _ index of the modeling segment.
  Thus, if \code{eff.id} is \code{1}, then the segment ID for 
  the second segment from \code{x} is \code{"1_2"}.
  
  The average condition values are calculated as a weighted average by distance, 
  and reported for the following:
  Beaufort, cloud cover, altitude in feet, aircraft speed in knots, jellyfish code, 
  and the percentage of the segment with each of haze, kelp, and red tide. 
  
  \code{\link[swfscMisc]{bearing}} and \code{\link[swfscMisc]{destination}} 
  are used to calculate the segment midpoints.
}
\keyword{internal}
