% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_sight.R
\name{airdas_sight}
\alias{airdas_sight}
\title{Aerial DAS sightings}
\usage{
airdas_sight(das.df)
}
\arguments{
\item{das.df}{data frame; processed aerial DAS data.. TODO}
}
\value{
Data frame with one row for each species for each sighting.
  Associated information (observer, species, etc.) is extracted 
  from \code{Data#} columns into as specified in Details.
  Other columns from \code{das.df} will be included in the output
}
\description{
Extract sightings and associated information from aerial DAS data
}
\details{
TODO. 
  TODO: make parameters for 'event.sight' and 'event.turtle'?
  Marine mammal sighting codes: "S" and "s"; turtle sighting code: "t"
  This function assumes the following 
  \tabular{llll}{
    \emph{Information} \tab \emph{Mammal sighting} \tab 
      \emph{Turtle sighting} \tab \emph{New column name}\cr
    Sighting number \tab Data1 \tab \tab sight_no\cr
    Observer \tab Data2 \tab Data1 \tab obs\cr
    Declination angle \tab Data3 \tab Data2 \tab angle_declination\cr
    Etc\cr
  }
}
\examples{
# TODO
# d <- do.call(rbind, lapply(list.files("../airdas/airDAS_files", full.names = TRUE), airdas_read))
# x <- airdas_process(d)
# x.sight <- airdas_sight(x)

}
