% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_sight.R
\name{airdas_sight}
\alias{airdas_sight}
\alias{airdas_sight.data.frame}
\alias{airdas_sight.airdas_df}
\title{Aerial DAS sightings}
\usage{
airdas_sight(x)

\method{airdas_sight}{data.frame}(x)

\method{airdas_sight}{airdas_df}(x)
}
\arguments{
\item{x}{\code{airdas_df} object; output from \code{\link{airdas_process}}, 
or a data frame that can be coerced to a \code{airdas_df} object}
}
\value{
Data frame with 1) the columns from \code{x}, excluding the 'Data#' columns,
  and 2) columns with sighting information (observer, species, group size, etc.) 
  extracted from 'Data#' columns as specified in Details.
  The data frame has one row for each sighting,
  or one row for each species of each sighting if 
  it is a multi-species sighting.
}
\description{
Extract sighting information from aerial DAS data
}
\details{
AirDAS events contain specific information in the 'Data#' columns,
  with the information depending on the event code and file type for that row.
  This function, which can handle multiple file types in \code{x} 
  (for instance, \code{x} could be combined PHOCOENA and TURTLE data) 
  extracts relevant data for sighting events, and returns a
  data frame with dedicated columns for each piece of sighting information.
  See \code{\link{airdas_format_pdf}} for more information about the expected
  events and event formats, depending on the file type.
  
  Abbreviations used in column names include: Gs = group size, Sp = species, 
  Mixed = mixed species (multi-species) sighting.#' 
  A 'standard sighting' ('SightStd' in output data frame) is a sighting 
  made by ObsL, ObsB, or ObsR (not the data recorder or pilot).#'   
  In addition, multi-species group sizes are rounded to 
  the nearest whole number using \code{round(, 0)}.
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

airdas_sight(y.proc)

}
