% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_sight.R
\name{airdas_sight}
\alias{airdas_sight}
\alias{airdas_sight.data.frame}
\alias{airdas_sight.airdas_df}
\title{Aerial DAS sightings}
\usage{
airdas_sight(x)

\method{airdas_sight}{data.frame}(x)

\method{airdas_sight}{airdas_df}(x)
}
\arguments{
\item{x}{\code{airdas_df} object; output from \code{\link{airdas_process}}, 
or a data frame that can be coerced to a \code{airdas_df} object}
}
\value{
Data frame with 1) the columns from \code{x}, excluding the 'Data#' columns,
  and 2) columns with sighting information
  (observer, species, etc.) extracted from 'Data#' columns as specified in Details.
  The data frame has one row for each sighting,
  or one row for each species of each sighting if it is a multi-species sighting.
}
\description{
Extract sightings and associated information from aerial DAS data
}
\details{
This function requires the following event codes: 
  
  Marine mammal sighting codes: "S" (first sighting) and "s" (resighting); 
  turtle sighting code: "t";
  additional sighting information: "1"
  
  For more info about the event codes, see 
  \url{https://github.com/smwoodman/swfscAirDAS/blob/master/inst/AirDAS_Format.pdf}
  
  A 'standard sighting' (SightStd in output data frame) means it was by ObsL, ObsB, or ObsR
  
  Multispecies group size is rounded to nearest integer using round(, 0)
  
  This function assumes the following 
  \tabular{llll}{
    \emph{Information} \tab \emph{Mammal sighting} \tab \emph{Turtle sighting} \tab \emph{New column name}\cr
    Sighting number   \tab Data1          \tab       \tab SightNo\cr
    Observer          \tab Data2          \tab Data1 \tab Obs\cr
    Declination angle \tab Data3          \tab Data2 \tab Angle\cr
    Group size (best estimate) \tab Data4 \tab NA    \tab Gs \cr
    Species code 1    \tab Data5          \tab Data3 \tab Sp \cr
    Species code 2    \tab Data6          \tab NA    \tab todo \cr
    Species code 3    \tab Data7          \tab NA    \tab todo \cr
    Turtle size (ft)  \tab NA             \tab Data4 \tab todo\cr
    Travel direction (deg) \tab NA        \tab Data5 \tab todo\cr
    Tail visible?     \tab NA             \tab Data6 \tab todo\cr
  }
}
\examples{
y <- system.file("airdas_sample.das", package = "swfscAirDAS")
y.proc <- airdas_process(y)

airdas_sight(y.proc)

}
