% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_spcodes.R
\name{airdas_spcodes}
\alias{airdas_spcodes}
\alias{airdas_spcodes_read}
\title{Read AirDAS SpCodes file}
\usage{
airdas_spcodes_read(file, skip = 3)
}
\arguments{
\item{file}{character; filename of .dat file from which to read accepted
species codes}

\item{skip}{integer; default is 3.
Number of lines to skip when reading SpCodes file.
See \code{\link[readr:read_fwf]{readr::read_fwf()}} for more details}
}
\value{
Data frame with three columns:
\itemize{
\item code: species code
\item common_name: species common name
( sci_name: species scientific name)
}
}
\description{
Read AirDAS SpCodes file
}
\details{
todo
}
\examples{
sp.codes.file <- system.file("extdata", "SpCodesAirDAS.dat", package = "swfscAirDAS")
airdas_spcodes_read(sp.codes.file)

}
