% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdas_spcodes.R
\name{airdas_spcodes}
\alias{airdas_spcodes}
\alias{airdas_spcodes_read}
\title{Read AirDAS SpCodes file}
\usage{
airdas_spcodes_read(file, skip = 3)
}
\arguments{
\item{file}{character; filename of .dat file from which to read accepted
species codes}

\item{skip}{integer; default is 3.
Number of lines to skip when reading SpCodes file.
See \code{\link[readr:read_fwf]{readr::read_fwf()}} for more details}
}
\value{
Data frame with three columns:
\itemize{
\item code: species code, columns 1 to 6
\item common_name: species common name, columns 10 to 42
\item sci_name: species scientific name, columns 43 until the end of the line
}
}
\description{
Read AirDAS SpCodes file
}
\details{
Provide a standardized function to read an aerial survey DAS SpCodes file.
Methods described in 'returns'
}
\examples{
sp.codes.file <- system.file(
  "extdata", "SpCodesAirDAS.dat", package = "swfscAirDAS"
)
if (interactive()) airdas_spcodes_read(sp.codes.file)

}
