% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_airdas_df.R
\name{as_airdas_df}
\alias{as_airdas_df}
\alias{as_airdas_df.airdas_df}
\alias{as_airdas_df.data.frame}
\title{Coerce object to a airdas_df object}
\usage{
as_airdas_df(x)

\method{as_airdas_df}{airdas_df}(x)

\method{as_airdas_df}{data.frame}(x)
}
\arguments{
\item{x}{A object to be coerced to class \code{airdas_df}}
}
\value{
An object of class `airdas_df`
}
\description{
Check if an object is of class \code{\link{airdas_df}}, or coerce it if possible.
}
\details{
Currently only data frames can be coerced to an object of class \code{\link{airdas_df}}.
  If the \code{x} does not have column names and classes as specified in \code{\link{airdas_df}},
  then the function returns an error message detailing the first column that does not
  meet the \code{\link{airdas_df}} requirements.
}
\seealso{
\code{\link{airdas_df-class}}
}
