% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfscAirDAS-package.R
\docType{package}
\name{swfscAirDAS-package}
\alias{swfscAirDAS-package}
\alias{swfscAirDAS}
\title{Southwest Fisheries Science Center Aerial Survey DAS}
\description{
Process and summarize aerial survey DAS data
}
\details{
This package contains functions designed for processing and analyzing 
aerial survey DAS data (AirDAS) generated using the TURTLEP program (or associated programs) 
by the Southwest Fisheries Science Center. 
Functionality includes reading AirDAS data into a data frame, 
processing this data (extracting state and condition information for each AirDAS event), 
and summarizing sighting and effort information.
}
\seealso{
\url{https://smwoodman.github.io/swfscAirDAS/}
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
